requirements.txt
fastapi==0.104.1
strawberry-graphql==0.195.1
SQLAlchemy==2.0.19
uvicorn[standard]==0.21.1
gunicorn==20.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
emails==0.6
psycopg2-binary==2.9.6
apscheduler==3.10.1
sqlalchemy_mptt==0.2.5
sqlalchemy_mixins==2.0.4
tenacity==8.2.2
pydantic[email]==2.4.2
pydantic_settings==2.0.3
python-multipart==0.0.6
alembic==1.11.1
asyncpg==0.27.0
broadcaster~=0.2.0
aiofiles==23.1.0
pandas==2.0.1
jinja2==3.1.2
celery==5.2.7
fpdf2==2.7.3
PyPDF2==3.0.1
reportlab==4.0.0
fhir.resources==7.0.0
httpx==0.24.0
importlib-metadata~=6.0.0
opentelemetry-instrumentation-fastapi==0.38b0
opentelemetry-exporter-otlp==1.17.0
opentelemetry-instrumentation-logging==0.38b0
opentelemetry-instrumentation-sqlalchemy==0.38b0
